Fixed gtest errors

+++ tests/argumentParser_test.cpp
@@ -36,7 +36,7 @@ void argumentParser_test(){
         EXPECT_FALSE(c->printDot());
         
         std::vector<std::string> signals = c->getSignals();
-        EXPECT_EQ(0, signals.size());
+        EXPECT_EQ((size_t)0, signals.size());
         delete c;
     }END;
 
@@ -52,11 +52,11 @@ void argumentParser_test(){
         EXPECT_FALSE(c->printDot());
         
         std::vector<std::string> signals = c->getSignals();
-        EXPECT_EQ(3, signals.size());
+        EXPECT_EQ((size_t)3, signals.size());
         EXPECT_STREQ("sig0", signals[0].c_str());
         EXPECT_STREQ("sig1", signals[1].c_str());
         EXPECT_STREQ("sig2", signals[2].c_str());
         delete c;
     }END;
-    #endif
+#endif
 }
